Split tasks into their own modules to facilitate testing
