+++ packages/cli/src/commands/setup/tailwind/__tests__/index.test.js
@@ -1,0 +1,1 @@
+global.__dirname = __dirname
