+++ packages/cli/src/commands/setup/tailwind/tailwind.js
@@ -1,12 +1,14 @@
-import fs from 'fs'
-import path from 'path'
-
 import Listr from 'listr'
-import execa from 'execa'
 import chalk from 'chalk'
 
 import c from 'src/lib/colors'
-import { getPaths, writeFile } from 'src/lib'
+import {
+  configurePostCSS,
+  installPackages,
+  yarnCheckFiles,
+  initTailwind,
+  addCSSImports,
+} from './tasks'
 
 export const command = 'tailwind'
 export const description = 'Setup tailwindcss and PostCSS'
@@ -25,179 +27,34 @@ export const builder = (yargs) => {
     })
 }
 
-const tailwindImportsAndNotes = [
-  '/**',
-  ' * START --- TAILWIND GENERATOR EDIT',
-  ' *',
-  ' * `yarn rw setup tailwind` placed these imports here',
-  " * to inject Tailwind's styles into your CSS.",
-  ' * For more information, see: https://tailwindcss.com/docs/installation#add-tailwind-to-your-css',
-  ' */',
-  '@import "tailwindcss/base";',
-  '@import "tailwindcss/components";',
-  '@import "tailwindcss/utilities";',
-  '/**',
-  ' * END --- TAILWIND GENERATOR EDIT',
-  ' */\n',
-]
-
-const INDEX_CSS_PATH = path.join(getPaths().web.src, 'index.css')
-
-const tailwindImportsExist = (indexCSS) => {
-  let content = indexCSS.toString()
-
-  const hasBaseImport = () => /@import "tailwindcss\/base"/.test(content)
-
-  const hasComponentsImport = () =>
-    /@import "tailwindcss\/components"/.test(content)
-
-  const hasUtilitiesImport = () =>
-    /@import "tailwindcss\/utilities"/.test(content)
-
-  return hasBaseImport() && hasComponentsImport() && hasUtilitiesImport()
-}
-
-const postCSSConfigExists = () => {
-  return fs.existsSync(getPaths().web.postcss)
-}
-
-export const handler = async ({ force, ui }) => {
-  const tasks = new Listr([
+const tasks = (args) =>
+  new Listr([
     {
       title: 'Installing packages...',
       task: () => {
         return new Listr([
           {
             title: 'Install postcss-loader, tailwindcss, and autoprefixer',
-            task: async () => {
-              /**
-               * Install postcss-loader, tailwindcss, and autoprefixer. Add TailwindUI if requested.
-               * RedwoodJS currently uses PostCSS v7; postcss-loader and autoprefixers pinned for compatibility
-               */
-              let packages = [
-                'postcss-loader@4.0.2',
-                'tailwindcss',
-                'autoprefixer@9.8.6',
-              ]
-
-              if (ui) {
-                packages.push('@tailwindcss/ui')
-              }
-
-              await execa('yarn', [
-                'workspace',
-                'web',
-                'add',
-                '-D',
-                ...packages,
-              ])
-            },
+            task: installPackages(args),
           },
           {
             title: 'Sync yarn.lock and node_modules',
-            task: async () => {
-              /**
-               * Sync yarn.lock file and node_modules folder.
-               * Refer https://github.com/redwoodjs/redwood/issues/1301 for more details.
-               */
-              await execa('yarn', ['install', '--check-files'])
-            },
+            task: yarnCheckFiles(args),
           },
         ])
       },
     },
     {
       title: 'Configuring PostCSS...',
-      task: () => {
-        /**
-         * Make web/config if it doesn't exist
-         * and write postcss.config.js there
-         */
-
-        /**
-         * Check if PostCSS config already exists.
-         * If it exists, throw an error.
-         */
-        if (!force && postCSSConfigExists()) {
-          throw new Error(
-            'PostCSS config already exists.\nUse --force to override existing config.'
-          )
-        } else {
-          return writeFile(
-            getPaths().web.postcss,
-            fs
-              .readFileSync(
-                path.resolve(
-                  __dirname,
-                  'templates',
-                  'postcss.config.js.template'
-                )
-              )
-              .toString(),
-            { overwriteExisting: force }
-          )
-        }
-      },
+      task: configurePostCSS(args),
     },
     {
       title: 'Initializing Tailwind CSS...',
-      task: async () => {
-        /**
-         * If it doesn't already exist,
-         * initialize tailwind and move tailwind.config.js to web/
-         */
-        const configExists = fs.existsSync(
-          path.join(getPaths().web.base, 'tailwind.config.js')
-        )
-
-        if (!force && configExists) {
-          throw new Error(
-            'Tailwindcss config already exists.\nUse --force to override existing config.'
-          )
-        } else {
-          await execa('yarn', ['tailwindcss', 'init'])
-
-          const config = fs.readFileSync('tailwind.config.js', 'utf-8')
-
-          // opt-in to upcoming changes
-          const uncommentFlags = (str) =>
-            str.replace(/\/{2} ([\w-]+: true)/g, '$1')
-
-          let newConfig = config.replace(/future.*purge/s, uncommentFlags)
-
-          // add TailwindUI plugin if requested
-          if (ui) {
-            newConfig = newConfig.replace(
-              /plugins:\W*\[\W*]/s,
-              "plugins: [require('@tailwindcss/ui')]"
-            )
-          }
-
-          fs.writeFileSync('tailwind.config.js', newConfig)
-
-          /**
-           * Later, when we can tell the vscode extension where to look for the config,
-           * we can put it in web/config/
-           */
-          await execa('mv', ['tailwind.config.js', 'web/'])
-        }
-      },
+      task: initTailwind(args),
     },
     {
       title: 'Adding imports to index.css...',
-      task: (_ctx, task) => {
-        /**
-         * Add tailwind imports and notes to the top of index.css
-         */
-        let indexCSS = fs.readFileSync(INDEX_CSS_PATH)
-
-        if (tailwindImportsExist(indexCSS)) {
-          task.skip('Imports already exist in index.css')
-        } else {
-          indexCSS = tailwindImportsAndNotes.join('\n') + indexCSS
-          fs.writeFileSync(INDEX_CSS_PATH, indexCSS)
-        }
-      },
+      task: addCSSImports(args),
     },
     {
       title: 'One more thing...',
@@ -214,8 +71,9 @@ export const handler = async ({ force, ui }) => {
     },
   ])
 
+export const handler = async (args) => {
   try {
-    await tasks.run()
+    await tasks(args).run()
   } catch (e) {
     console.log(c.error(e.message))
   }
