+++ packages/cli/src/commands/setup/tailwind/tasks/index.js
@@ -1,0 +1,5 @@
+export { default as installPackages } from './installPackages'
+export { default as yarnCheckFiles } from './yarnCheckFiles'
+export { default as configurePostCSS } from './configurePostCSS'
+export { default as initTailwind } from './initTailwind'
+export { default as addCSSImports } from './addCSSImports'
